+++ lib/flutter_callkit_incoming.dart
@@ -58,6 +58,13 @@ class FlutterCallkitIncoming {
     await _channel.invokeMethod("endAllCalls");
   }
 
+  /// Get active calls.
+  /// On iOS: return active calls from Callkit.
+  /// On Android: only return last call
+  static Future activeCalls() async {
+    return await _channel.invokeMethod("activeCalls");
+  }
+
   static CallEvent? _receiveCallEvent(dynamic data) {
     var event = "";
     dynamic body = {};