+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.0+4
+
+* Update README.md.
+* Add func `activeCalls()` Get active calls
+* Remove notification when click action `Call back` (Android).
+* Bugs `no activation of the audio session` (iOS)
+* Fixed some bugs.
+
 ## 1.0.0+3
 
 * Update README.md.